Move functions to io_lib Signed-off-by: Chipmuenk <mail@chipmuenk.de>
